add admin gate in edit&destroy functions

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -46,6 +47,10 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users')){
+            return redirect()->route('user.index');
+
+        }
         $roles=Role::all();
         return view('admin.users.edit',[
             'user'=>$user,
@@ -63,6 +68,10 @@ class UserController extends Controller
     
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users')){
+            return redirect()->route('user.index');
+
+        }
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('user.index');
